fix: improve handling of linked lists
walking through the linked list is inefficient and might lead to
errors. this fix adds to the end of the list directly instead.
as a result, this closes #11, since double trigger events seem to be
resolved. also fixes a leak where monome devices were not properly
cleaned up on (normal) exit.